+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 - Fix: The field publicationURL of a pubblication is now rendered as an 'a' tag
   in the template (#16555)
   [arsenico13]
+- Fix: Now the vocabulary for the authors of a pubblication is correct and
+  filters the names while typing in the edit view (2.a of #16555)
+  [arsenico13]
 
 
 2.0.3 (2018-07-05)